requirements.txt
pyopengl==3.1.7
glfw==2.6.2
pyrr==0.10.3